requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary

python-dotenv

email-validator
passlib[bcrypt]
bcrypt==4.0.1
